Feat/finalize stripe payments

+++ apps/api/src/download/download.controller.ts
@@ -1,18 +1,24 @@
 import {
     Controller,
+    ForbiddenException,
     Get,
+    GoneException,
     HttpException,
     HttpStatus,
     Ip,
+    Logger,
     NotFoundException,
     Post,
     Query,
     Res,
     StreamableFile,
+    UnauthorizedException,
 } from "@nestjs/common";
 import {
     ApiBadGatewayResponse,
     ApiCreatedResponse,
+    ApiForbiddenResponse,
+    ApiGoneResponse,
     ApiNotFoundResponse,
     ApiOkResponse,
     ApiProduces,
@@ -34,6 +40,8 @@ import { join } from "path";
 import { DownloadsItemDto } from "@better-bedrock/constants/downloads.dto";
 import { COMMUNITY_LIST } from "src/content/constants/content-community";
 import { SIDE_PROJECTS_LIST } from "src/content/constants/content-side-projects";
+import { VoucherService } from "src/voucher/voucher.service";
+import { SkipThrottle } from "@nestjs/throttler";
 
 @ApiTags("download")
 @Controller("download")
@@ -44,18 +52,26 @@ export class DownloadController {
         private readonly http: HttpService,
         private downloadService: DownloadService,
         private analyticsService: AnalyticsService,
+        private readonly voucherService: VoucherService,
     ) {}
 
     @Get()
+    @SkipThrottle()
     @ApiOkResponse({
         description: "Binary file stream",
         schema: { type: "string", format: "binary" },
     })
     @ApiProduces("application/octet-stream")
-    @ApiUnauthorizedResponse({ description: "Not verified for download." })
-    @ApiNotFoundResponse({ description: "File not found or does not exist on server." })
+    @ApiUnauthorizedResponse({ description: "Not verified for download" })
+    @ApiNotFoundResponse({ description: "File not found or does not exist on the server" })
     async download(@Ip() ip, @Res({ passthrough: true }) res: Response): Promise<StreamableFile> {
         const download = await this.downloadService.download({ ipAddress: ip });
+        //TODO: Add try check
+
+        if (!download) {
+            throw new NotFoundException(`Download for your device could not be found.`);
+        }
+
         if (!download?.verified) {
             throw new HttpException(
                 "You have not verified your download yet.",
@@ -115,61 +131,99 @@ export class DownloadController {
     }
 
     @Post("verify")
-    @ApiQuery({
-        name: "hash",
-        type: "string",
-        description: "Linkvertise anti-bypass hash",
-        required: true,
-    })
-    @ApiOkResponse({ type: DownloadsItemDto, description: "Download verified successfully." })
+    @ApiOkResponse({ type: DownloadsItemDto, description: "Download verified successfully" })
     @ApiNotFoundResponse({
-        description: "Download record not found for this IP or file not found.",
+        description: "Download record not found for this IP or file not found",
     })
-    @ApiBadGatewayResponse({ description: "Failed to verify with Linkvertise gateway." })
-    @ApiServiceUnavailableResponse({ description: "Linkvertise service unavailable." })
+    @ApiGoneResponse({ description: "The voucher has either expired or already been used" })
+    @ApiForbiddenResponse({
+        description: "This voucher allows you to download only better bedrock content",
+    })
+    @ApiUnauthorizedResponse({ description: "The voucher does not exist" })
+    @ApiBadGatewayResponse({ description: "Failed to verify with Linkvertise gateway" })
+    @ApiServiceUnavailableResponse({ description: "Linkvertise service unavailable" })
     async verify(
         @Ip() ip: string,
         @Query() query: VerifyDownloadDto,
     ): Promise<DownloadsItemDto | undefined> {
         const download = await this.downloadService.download({ ipAddress: ip });
-
+        const voucher = await this.voucherService.getVoucher({ code: query.code });
         if (!download) {
             throw new NotFoundException(`Download for your device could not be found.`);
         }
 
+        const file = this.findDownloadItemById(download.file);
+
         const url = `https://publisher.linkvertise.com/api/v1/anti_bypassing?token=${this.linkvertiseApiKey}&hash=${query.hash}`;
 
         if (!download.verified) {
-            try {
-                const response$ = this.http.post(
-                    url,
-                    {}, // no body
-                    {
-                        headers: {
-                            "Access-Control-Allow-Origin": "*",
-                            "Access-Control-Allow-Methods": "POST",
-                            "Access-Control-Allow-Headers": "Content-Type, Authorization",
-                        },
-                    },
-                );
-                const { data, status } = await lastValueFrom(response$);
+            if (query.code && !voucher) {
+                throw new UnauthorizedException("The voucher does not exist");
+            }
 
-                if (status !== 200 || !data.status) {
-                    throw new HttpException(
-                        "Failed to verify with Linkvertise gateway.",
-                        HttpStatus.BAD_GATEWAY,
+            if (voucher) {
+                if (
+                    voucher.downloadCount >= voucher.maxDownloads ||
+                    voucher.expiresAt.getTime() <= Date.now()
+                ) {
+                    throw new GoneException("The voucher has either expired or already been used.");
+                }
+
+                if (
+                    voucher.betterBedrockContentOnly &&
+                    DOWNLOADS_LIST.flatMap((map) => map.items).find(
+                        (item) => item.downloadId === file!.downloadId,
+                    )
+                ) {
+                    throw new ForbiddenException(
+                        "This voucher allows you to download only better bedrock content",
                     );
                 }
-            } catch (err) {
-                if (err instanceof HttpException) {
-                    throw err;
-                }
-
-                throw new HttpException(
-                    "Linkvertise service unavailable.",
-                    HttpStatus.SERVICE_UNAVAILABLE,
-                );
             }
+
+            if (!voucher) {
+                try {
+                    const response$ = this.http.post(
+                        url,
+                        {}, // no body
+                        {
+                            headers: {
+                                "Access-Control-Allow-Origin": "*",
+                                "Access-Control-Allow-Methods": "POST",
+                                "Access-Control-Allow-Headers": "Content-Type, Authorization",
+                            },
+                        },
+                    );
+                    const { data, status } = await lastValueFrom(response$);
+
+                    if (status !== 200 || !data.status) {
+                        throw new HttpException(
+                            "Failed to verify with Linkvertise gateway.",
+                            HttpStatus.BAD_GATEWAY,
+                        );
+                    }
+                } catch (err) {
+                    if (err instanceof HttpException) {
+                        throw err;
+                    }
+
+                    throw new HttpException(
+                        "Linkvertise service unavailable.",
+                        HttpStatus.SERVICE_UNAVAILABLE,
+                    );
+                }
+            }
+
+            Logger.error("Download verified: ");
+
+            await this.voucherService.updateVoucher({
+                where: { code: query.code },
+                data: {
+                    downloadCount: {
+                        increment: 1,
+                    },
+                },
+            });
 
             await this.downloadService.updateDownload({
                 where: { ipAddress: ip },
@@ -190,20 +244,36 @@ export class DownloadController {
     @ApiCreatedResponse({ description: "Download record created." })
     @ApiNotFoundResponse({ description: "Requested file not found." })
     async generate(@Ip() ip, @Query() query: GenerateDownloadDto) {
-        this.findDownloadItemById(query.file);
+        try {
+            this.findDownloadItemById(query.file);
+            const download = await this.downloadService.download({ ipAddress: ip });
 
-        await this.downloadService.deleteDownload({ ipAddress: ip });
-        await this.downloadService.createDownload({
-            ipAddress: ip,
-            file: query.file,
-        });
+            if (download != null) {
+                try {
+                    await this.downloadService.deleteDownload({ ipAddress: ip });
+                } catch (_) {
+                    /* empty */
+                }
+            }
 
-        // Update analytics
-        const analytics = await this.analyticsService.analytics();
-        if (analytics) {
-            await this.analyticsService.updateAnalytics({
-                totalDownloads: analytics.totalDownloads++,
-            });
+            try {
+                await this.downloadService.createDownload({
+                    ipAddress: ip,
+                    file: query.file,
+                });
+            } catch (_) {
+                /* empty */
+            }
+
+            // Update analytics
+            const analytics = await this.analyticsService.analytics();
+            if (analytics) {
+                await this.analyticsService.updateAnalytics({
+                    totalDownloads: analytics.totalDownloads++,
+                });
+            }
+        } catch (err) {
+            throw new HttpException("Error occured during download generation: " + err, 500);
         }
     }
